Added dialog to allow user to attempt opening a file as plain text if no app is registered for the deduced mime type from it's extension

+++ app/src/main/java/org/linphone/utils/LinphoneUtils.kt
@@ -79,7 +79,7 @@ class LinphoneUtils {
         fun deleteFilesAttachedToChatMessage(chatMessage: ChatMessage) {
             for (content in chatMessage.contents) {
                 val filePath = content.filePath
-                if (filePath != null) {
+                if (filePath != null && filePath.isNotEmpty()) {
                     Log.i("[Linphone Utils] Deleting file $filePath")
                     FileUtils.deleteFile(filePath)
                 }
